+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add methods to client for using in java
+- Parse error response with alerts from Ground service
+
 ## [3.0.0] - 2023-03-24
 
 ### Added